refactor(www): separate button tokens from semantic color tokens file
