+++ tokens/tools/tokens-build.js
@@ -1,6 +1,7 @@
 const generateTokenJSON = require("./generateTokens");
 
 const border = require("../src/core/border");
+const button = require("../src/core/button");
 const colors = require("../src/core/colors");
 const font = require("../src/core/font");
 const lineHeight = require("../src/core/lineHeight");
@@ -12,6 +13,7 @@ const zIndex = require("../src/core/zIndex");
 (async function() {
   try {
     await generateTokenJSON('border', border);
+    await generateTokenJSON('button', button);
     await generateTokenJSON('colors', colors);
     await generateTokenJSON('font', font);
     await generateTokenJSON('lineHeight', lineHeight);
